fix: 修正 clippy 警告（不必要 cast、多餘 closure、複雜型別）
- expiry_config.rs: 移除 `cutoff as i16`（cutoff 已是 i16）
- expiry_monthly.rs: 以 Self::snapshot_key 取代冗餘 closure
- expiry_monthly.rs: 提取 SnapshotRow 型別別名解決 type_complexity

+++ backend/src/services/notification/expiry_config.rs
@@ -36,7 +36,7 @@ impl NotificationService {
     ) -> Result<ExpiryNotificationConfig, AppError> {
         if let Some(threshold) = request.monthly_threshold_days {
             let cutoff = request.cutoff_days.unwrap_or(90);
-            if threshold > cutoff as i16 {
+            if threshold > cutoff {
                 return Err(AppError::BadRequest(
                     "monthly_threshold_days 不能大於 cutoff_days".to_string(),
                 ));
